+++ lib/clixon/clixon_string.h
@@ -87,6 +87,7 @@ int uri_percent_decode(char *enc, char **str);
 const char *clicon_int2str(const map_str2int *mstab, int i);
 int clicon_str2int(const map_str2int *mstab, char *str);
 int nodeid_split(char *nodeid,	char **prefix, char **id);
+char *clixon_trim(char *str);
 #ifndef HAVE_STRNDUP
 char *clicon_strndup (const char *, size_t);
 #endif /* ! HAVE_STRNDUP */