+++ lib/src/clixon_xml_map.c
@@ -2414,7 +2414,7 @@ yang_enum_int_value(cxobj   *node,
     if ((ytype = yang_find((yang_node *)ys, Y_TYPE, NULL)) == NULL)
 	goto done;
     if (yang_type_resolve(ys, ytype, &yrestype, 
-			  NULL, NULL, NULL, NULL, NULL) < 0)
+			  NULL, NULL, NULL, NULL) < 0)
 	goto done;
     if (yrestype==NULL || strcmp(yrestype->ys_argument, "enumeration"))
 	goto done;