+++ lib/src/clixon_yang_type.c
@@ -115,8 +115,7 @@ int
 yang_type_cache_set(yang_type_cache **ycache0,
 		    yang_stmt        *resolved,
 		    int               options, 
-		    cg_var           *mincv, 
-		    cg_var           *maxcv, 
+		    cvec             *cvv, 
 		    char             *pattern,
 		    uint8_t           fraction)
 {
@@ -132,13 +131,11 @@ yang_type_cache_set(yang_type_cache **ycache0,
     *ycache0 = ycache;
     ycache->yc_resolved  = resolved;
     ycache->yc_options  = options;
-    if (mincv && (ycache->yc_mincv  = cv_dup(mincv)) == NULL){
-	clicon_err(OE_UNIX, errno, "cv_dup");
-	goto done;
-    }
-    if (maxcv && (ycache->yc_maxcv  = cv_dup(maxcv)) == NULL){
-	clicon_err(OE_UNIX, errno, "cv_dup");
-	goto done;
+    if (cvv){
+	if ((ycache->yc_cvv = cvec_dup(cvv)) == NULL){
+	    clicon_err(OE_UNIX, errno, "cvec_dup");
+	    goto done;
+	}
     }
     if (pattern && (ycache->yc_pattern  = strdup(pattern)) == NULL){
 	clicon_err(OE_UNIX, errno, "strdup");
@@ -155,8 +152,7 @@ int
 yang_type_cache_get(yang_type_cache *ycache,
 		    yang_stmt      **resolved,
 		    int             *options, 
-		    cg_var         **mincv, 
-		    cg_var         **maxcv, 
+		    cvec           **cvv, 
 		    char           **pattern,
 		    uint8_t         *fraction)
 {
@@ -164,10 +160,8 @@ yang_type_cache_get(yang_type_cache *ycache,
 	*resolved = ycache->yc_resolved;
     if (options)
 	*options  = ycache->yc_options;
-    if (mincv)
-	*mincv    = ycache->yc_mincv;
-    if (maxcv)
-	*maxcv    = ycache->yc_maxcv;
+    if (cvv)
+	*cvv    = ycache->yc_cvv;
     if (pattern)
 	*pattern  = ycache->yc_pattern;
     if (fraction)
@@ -181,14 +175,13 @@ yang_type_cache_cp(yang_type_cache **ycnew,
 {
     int        retval = -1;
     int        options;
-    cg_var    *mincv;
-    cg_var    *maxcv;
+    cvec      *cvv;
     char      *pattern;
     uint8_t    fraction;
     yang_stmt *resolved;
 
-    yang_type_cache_get(ycold, &resolved, &options, &mincv, &maxcv, &pattern, &fraction);
-    if (yang_type_cache_set(ycnew, resolved, options, mincv, maxcv, pattern, fraction) < 0)
+    yang_type_cache_get(ycold, &resolved, &options, &cvv, &pattern, &fraction);
+    if (yang_type_cache_set(ycnew, resolved, options, cvv, pattern, fraction) < 0)
 	goto done;
     retval = 0;
  done:
@@ -198,10 +191,8 @@ yang_type_cache_cp(yang_type_cache **ycnew,
 int
 yang_type_cache_free(yang_type_cache *ycache)
 {
-    if (ycache->yc_mincv)
-	cv_free(ycache->yc_mincv);
-    if (ycache->yc_maxcv)
-	cv_free(ycache->yc_maxcv);
+    if (ycache->yc_cvv)
+	cvec_free(ycache->yc_cvv);
     if (ycache->yc_pattern)
 	free(ycache->yc_pattern);
     free(ycache);
@@ -220,8 +211,7 @@ ys_resolve_type(yang_stmt *ys,
 {
     int               retval = -1;
     int               options = 0x0;
-    cg_var           *mincv = NULL;
-    cg_var           *maxcv = NULL;
+    cvec             *cvv = NULL;
     char             *pattern = NULL;
     uint8_t           fraction = 0;
     yang_stmt        *resolved = NULL;
@@ -231,12 +221,12 @@ ys_resolve_type(yang_stmt *ys,
      * Note that the resolved type could be ys itself.
      */
     if (yang_type_resolve((yang_stmt*)ys->ys_parent, ys, &resolved,
-			  &options, &mincv, &maxcv, &pattern, &fraction) < 0)
+			  &options, &cvv, &pattern, &fraction) < 0)
 	goto done;
 
     /* Cache the resolve locally */
     if (yang_type_cache_set(&ys->ys_typecache, 
-			    resolved, options, mincv, maxcv, pattern, fraction) < 0)
+			    resolved, options, cvv, pattern, fraction) < 0)
 	goto done;
     retval = 0;
  done:
@@ -388,132 +378,121 @@ static int
 cv_validate1(cg_var      *cv,
 	     enum cv_type cvtype, 
 	     int          options, 
-	     cg_var      *range_min,
-	     cg_var      *range_max, 
+	     cvec        *cvv,
 	     char        *pattern,
 	     yang_stmt   *yrestype,
 	     char        *restype,
 	     char       **reason)
 {
     int             retval = 1; /* OK */
+    cg_var         *cv1;
+    cg_var         *cv2;
     int             retval2;
     yang_stmt      *yi = NULL;
-    unsigned int    u = 0;
-    int             i = 0;
     char           *str;
     int             found;
     char          **vec = NULL;
     int             nvec;
     char           *v;
+    uint64_t        uu = 0;
+    int64_t         ii = 0;
+    int             i;
+    int             ret;
 
     if (reason && *reason){
 	free(*reason);
 	*reason = NULL;
     }
+    /* check options first for length and range */
+    if ((options & YANG_OPTIONS_RANGE) != 0 ||
+	(options & YANG_OPTIONS_LENGTH) != 0){
+	i = 0;
+	while (i<cvec_len(cvv)){
+	    cv1 = cvec_i(cvv, i++);
+	    if (strcmp(cv_name_get(cv1),"range_min") == 0){
+		if (i<cvec_len(cvv) &&
+		    (cv2 = cvec_i(cvv, i)) != NULL &&
+		    strcmp(cv_name_get(cv2),"range_max") == 0){
+		    i++;
+		}
+		else
+		    cv2 = cv1;
+		ret = 0;
+		switch (cvtype){
+		case CGV_INT8:
+		    ii =  cv_int8_get(cv);
+		    ret = range_check(ii, cv1, cv2, int8);
+		    break;
+		case CGV_INT16:
+		    ii =  cv_int16_get(cv);
+		    ret = range_check(ii, cv1, cv2, int16);
+		    break;
+		case CGV_INT32:
+		    ii =  cv_int32_get(cv);
+		    ret = range_check(ii, cv1, cv2, int32);
+		    break;
+		case CGV_DEC64: /* XXX look at fraction-digit? */
+		case CGV_INT64:
+		    ii =  cv_int64_get(cv);
+		    ret = range_check(ii, cv1, cv2, int64);
+		    break;
+		default:
+		    break;
+		}
+		if (ret){
+		    if (reason)
+			*reason = cligen_reason("Number out of range: %" PRId64, ii);
+		    goto fail;
+		    break;
+		}
+		ret = 0;
+		switch (cvtype){
+		case CGV_UINT8:
+		    uu =  cv_uint8_get(cv);
+		    ret = range_check(uu, cv1, cv2, uint8);
+		    break;
+		case CGV_UINT16:
+		    uu =  cv_uint16_get(cv);
+		    ret = range_check(uu, cv1, cv2, uint16);
+		    break;
+		case CGV_UINT32:
+		    uu =  cv_uint32_get(cv);
+		    ret = range_check(uu, cv1, cv2, uint32);
+		    break;
+		case CGV_UINT64:
+		    uu =  cv_uint32_get(cv);
+		    ret = range_check(uu, cv1, cv2, uint64);
+		    break;
+		default:
+		    break;
+		}
+		if (ret){
+		    if (reason)
+			*reason = cligen_reason("Number out of range: %" PRIu64, uu);
+		    goto fail;
+		    break;
+		}
+		ret = 0;
+		switch (cvtype){
+		case CGV_STRING:
+		case CGV_REST:
+		    if ((str = cv_string_get(cv)) == NULL)
+			break;
+		    uu = strlen(str);
+		    if (range_check(uu, cv1, cv2, uint64)){
+			if (reason)
+			    *reason = cligen_reason("string length out of range: %" PRIu64, uu);
+			goto fail;
+		    }
+		    break;
+		default:
+		    break;
+		}
+	    }
+	}
+    }
+    /* then check options for others */
     switch (cvtype){
-    case CGV_INT8:
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    i = cv_int8_get(cv);
-	    if (range_check(i, range_min, range_max, int8)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %d", i);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    case CGV_INT16:
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    i = cv_int16_get(cv);
-	    if (range_check(i, range_min, range_max, int16)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %d", i);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    case CGV_INT32:
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    i = cv_int32_get(cv);
-	    if (range_check(i, range_min, range_max, int32)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %d", i);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    case CGV_INT64:{
-	int64_t i64;
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    i64 = cv_int64_get(cv);
-	    if (range_check(i, range_min, range_max, int64)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %" PRId64, i64);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    }
-    case CGV_UINT8:
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    u = cv_uint8_get(cv);
-	    if (range_check(u, range_min, range_max, uint8)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %u", u);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    case CGV_UINT16:
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    u = cv_uint16_get(cv);
-	    if (range_check(u, range_min, range_max, uint16)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %u", u);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    case CGV_UINT32:
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    u = cv_uint32_get(cv);
-	    if (range_check(u, range_min, range_max, uint32)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %u", u);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    case CGV_UINT64:{
-	uint64_t u64;
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    u64 = cv_uint64_get(cv);
-	    if (range_check(u, range_min, range_max, uint64)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %" PRIu64, u64);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
-    }
-    case CGV_DEC64:
-	if ((options & YANG_OPTIONS_RANGE) != 0){
-	    i = cv_int64_get(cv);
-	    if (range_check(i, range_min, range_max, int64)){
-		if (reason)
-		    *reason = cligen_reason("Number out of range: %d", i);
-		retval = 0;
-		break;
-	    }
-	}
-	break;
     case CGV_STRING:
     case CGV_REST:
 	str = cv_string_get(cv);
@@ -531,8 +510,7 @@ cv_validate1(cg_var      *cv,
 		if (!found){
 		    if (reason)
 			*reason = cligen_reason("'%s' does not match enumeration", str);
-		    retval = 0;
-		    break;
+		    goto fail;
 		}
 	    }
 	    if (strcmp(restype, "bits") == 0){
@@ -557,19 +535,10 @@ cv_validate1(cg_var      *cv,
 		    if (!found){
 			if (reason)
 			    *reason = cligen_reason("'%s' does not match enumeration", v);
-			retval = 0;
+			goto fail;
 			break;
 		    }
 		}
-	    }
-	}
-	if ((options & YANG_OPTIONS_LENGTH) != 0){
-	    u = strlen(str);
-	    if (range_check(u, range_min, range_max, uint64)){
-		if (reason)
-		    *reason = cligen_reason("string length out of range: %u", u);
-		retval = 0;
-		break;
 	    }
 	}
 	if ((options & YANG_OPTIONS_PATTERN) != 0){
@@ -581,7 +550,7 @@ cv_validate1(cg_var      *cv,
 		if (reason)
 		    *reason = cligen_reason("regexp match fail: \"%s\" does not match %s",
 					    str, pattern);
-		retval = 0;
+		goto fail;
 		break;
 	    }
 	}
@@ -591,27 +560,19 @@ cv_validate1(cg_var      *cv,
 	retval = 0;
 	if (reason)
 	    *reason = cligen_reason("Invalid cv");
-	retval = 0;
+	goto fail;
 	break;
-    case CGV_BOOL:
-    case CGV_INTERFACE:
-    case CGV_IPV4ADDR: 
-    case CGV_IPV6ADDR: 
-    case CGV_IPV4PFX: 
-    case CGV_IPV6PFX: 
-    case CGV_MACADDR:
-    case CGV_URL: 
-    case CGV_UUID: 
-    case CGV_TIME: 
-    case CGV_EMPTY:  /* XXX */
+    default:
 	break;
     }
-    if (reason && *reason)
-	assert(retval == 0); /* validation failed with error reason */
+    retval = 1; /* validation OK */
  done:
     if (vec)
 	free(vec);
     return retval;
+ fail:
+    retval = 0; /* validation failed */
+    goto done;
 }
 
 /* Forward */
@@ -633,16 +594,14 @@ ys_cv_validate_union_one(yang_stmt *ys,
     int          retval = -1;
     yang_stmt   *yrt;      /* union subtype */
     int          options = 0;
-    cg_var      *range_min = NULL; 
-    cg_var      *range_max = NULL; 
+    cvec        *cvv = NULL;
     char        *pattern = NULL;
     uint8_t      fraction = 0; 
     char        *restype;
     enum cv_type cvtype;
     cg_var      *cvt=NULL;
 
-    if (yang_type_resolve(ys, yt, &yrt, 
-			  &options, &range_min, &range_max, &pattern, 
+    if (yang_type_resolve(ys, yt, &yrt, &options, &cvv, &pattern,
 			  &fraction) < 0)
 	goto done;
     restype = yrt?yrt->ys_argument:NULL;
@@ -664,7 +623,7 @@ ys_cv_validate_union_one(yang_stmt *ys,
 	}
 	if (retval == 0)
 	    goto done;
-	if ((retval = cv_validate1(cvt, cvtype, options, range_min, range_max, 
+	if ((retval = cv_validate1(cvt, cvtype, options, cvv, 
 				   pattern, yrt, restype, reason)) < 0)
 	    goto done;
     }
@@ -736,8 +695,7 @@ ys_cv_validate(cg_var    *cv,
     int             retval = -1; 
     cg_var         *ycv;        /* cv of yang-statement */  
     int             options = 0;
-    cg_var         *range_min = NULL; 
-    cg_var         *range_max = NULL; 
+    cvec           *cvv = NULL;
     char           *pattern = NULL;
     enum cv_type    cvtype;
     char           *type;  /* orig type */
@@ -756,8 +714,7 @@ ys_cv_validate(cg_var    *cv,
     }
     ycv = ys->ys_cv;
     if (yang_type_get(ys, &type, &yrestype, 
-		      &options, &range_min, &range_max, &pattern,
-		      &fraction) < 0)
+		      &options, &cvv, &pattern, &fraction) < 0)
 	goto done;
     restype = yrestype?yrestype->ys_argument:NULL;
     if (clicon_type2cv(type, restype, &cvtype) < 0)
@@ -782,7 +739,7 @@ ys_cv_validate(cg_var    *cv,
 	retval = retval2; /* invalid (0) with latest reason or valid 1 */
     }
     else
-	if ((retval = cv_validate1(cv, cvtype, options, range_min, range_max, pattern,
+	if ((retval = cv_validate1(cv, cvtype, options, cvv, pattern,
 				   yrestype, restype, reason)) < 0)
 	    goto done;
   done:
@@ -889,7 +846,17 @@ yang_find_identity(yang_stmt *ys,
     return yid;
 }
 
-/*
+/*! Resolve type restrictions, return contraining parameters
+ * @param[in]  yrange    Yang type range restriction if any
+ * @param[in]  ylength   Yang type length restriction if any
+ * @param[in]  ypattern  Yang type pattern restriction if any
+ * @param[in]  yfraction Yang type fraction restriction if any
+ * @param[out] options   Pointer to flags field of optional values. optional
+ * @param[out] cvv       Pointer to cvec with min range or length. 
+ *                       If options&YANG_OPTIONS_RANGE or YANG_OPTIONS_LENGTH
+ * @param[out] pattern   Pointer to static string of yang string pattern. optional
+ * @param[out] fraction  For decimal64, how many digits after period
+ * @retval      0        OK. 
  */
 static int
 resolve_restrictions(yang_stmt   *yrange,
@@ -897,19 +864,16 @@ resolve_restrictions(yang_stmt   *yrange,
 		     yang_stmt   *ypattern,
 		     yang_stmt   *yfraction,
 		     int         *options, 
-		     cg_var     **mincv, 
-		     cg_var     **maxcv, 
+		     cvec       **cvv, 
 		     char       **pattern,
 		     uint8_t     *fraction)
 {
-    if (options && mincv && maxcv && yrange != NULL){
-	*mincv = cvec_find(yrange->ys_cvec, "range_min");
-	*maxcv = cvec_find(yrange->ys_cvec, "range_max");
+    if (options && cvv && yrange != NULL){
+	*cvv = yrange->ys_cvec;
 	*options  |= YANG_OPTIONS_RANGE;
     }
-    if (options && mincv && maxcv && ylength != NULL){
-	*mincv = cvec_find(ylength->ys_cvec, "range_min"); /* XXX fel typ */
-	*maxcv = cvec_find(ylength->ys_cvec, "range_max");
+    if (options && cvv && ylength != NULL){
+	*cvv = ylength->ys_cvec;
 	*options  |= YANG_OPTIONS_LENGTH;
     }
     if (options && pattern && ypattern != NULL){
@@ -928,8 +892,8 @@ resolve_restrictions(yang_stmt   *yrange,
  * @param[in]  ytype    yang-stmt object containing currently resolving type
  * @param[out] yrestype resolved type. return built-in type or NULL. mandatory
  * @param[out] options  pointer to flags field of optional values. optional
- * @param[out] mincv    pointer to cv with min range or length. If options&YANG_OPTIONS_RANGE
- * @param[out] maxcv    pointer to cv with max range or length. If options&YANG_OPTIONS_RANGE
+ * @param[out] cvv      pointer to cvec with min range or length. 
+ *                      If options&YANG_OPTIONS_RANGE or YANG_OPTIONS_LENGTH
  * @param[out] pattern  pointer to static string of yang string pattern. optional
  * @param[out] fraction for decimal64, how many digits after period
  * @retval      0        OK. Note yrestype may still be NULL.
@@ -946,8 +910,7 @@ yang_type_resolve(yang_stmt   *ys,
 		  yang_stmt   *ytype, 
 		  yang_stmt  **yrestype, 
 		  int         *options, 
-		  cg_var     **mincv, 
-		  cg_var     **maxcv, 
+		  cvec       **cvv, 
 		  char       **pattern,
 		  uint8_t     *fraction)
 {
@@ -970,8 +933,8 @@ yang_type_resolve(yang_stmt   *ys,
     prefix    = yarg_prefix(ytype); /* And this its prefix */
     /* Cache does not work for eg string length 32? */
     if (!yang_builtin(type) && ytype->ys_typecache != NULL){
-	if (yang_type_cache_get(ytype->ys_typecache, 
-				yrestype, options, mincv, maxcv, pattern, fraction) < 0)
+	if (yang_type_cache_get(ytype->ys_typecache, yrestype,
+				options, cvv, pattern, fraction) < 0)
 	    goto done;
 	goto ok;
     }
@@ -984,7 +947,7 @@ yang_type_resolve(yang_stmt   *ys,
     if (prefix == NULL && yang_builtin(type)){
 	*yrestype = ytype; 
 	resolve_restrictions(yrange, ylength, ypattern, yfraction, options, 
-			     mincv, maxcv, pattern, fraction);
+			     cvv, pattern, fraction);
 	goto ok;
     }
 
@@ -1022,13 +985,12 @@ yang_type_resolve(yang_stmt   *ys,
 	}
 	/* recursively resolve this new type */
 	if (yang_type_resolve(ys, rytype, yrestype, 
-			      options, mincv, maxcv, pattern, fraction) < 0)
+			      options, cvv, pattern, fraction) < 0)
 	    goto done;
 	/* overwrites the resolved if any */
-	resolve_restrictions(yrange, ylength, ypattern, yfraction, 
-			     options, mincv, maxcv, pattern, fraction);
+	resolve_restrictions(yrange, ylength, ypattern, yfraction, options,
+			     cvv, pattern, fraction);
     }
-
   ok:
     retval = 0;
   done:
@@ -1042,11 +1004,11 @@ yang_type_resolve(yang_stmt   *ys,
  * @code
  *   yang_stmt    *yrestype;
  *   int           options;
- *   int64_t       min, max;
+ *   cvec         *cvv = NULL;
  *   char         *pattern;
  *   uint8_t       fraction;
  *
- *   if (yang_type_get(ys, &type, &yrestype, &options, &min, &max, &pattern, &fraction) < 0)
+ *   if (yang_type_get(ys, &type, &yrestype, &options, &cvv, &pattern, &fraction) < 0)
  *      goto err;
  *   if (yrestype == NULL) # unresolved
  *      goto err;
@@ -1078,8 +1040,7 @@ yang_type_get(yang_stmt    *ys,
 	      char        **origtype, 
 	      yang_stmt   **yrestype, 
 	      int          *options, 
-	      cg_var      **mincv, 
-	      cg_var      **maxcv, 
+	      cvec        **cvv, 
 	      char        **pattern,
 	      uint8_t      *fraction
     )
@@ -1100,7 +1061,7 @@ yang_type_get(yang_stmt    *ys,
     if (origtype)
 	*origtype = type;
     if (yang_type_resolve(ys, ytype, yrestype, 
-			  options, mincv, maxcv, pattern, fraction) < 0)
+			  options, cvv, pattern, fraction) < 0)
 	goto done;
     clicon_debug(3, "%s: %s %s->%s", __FUNCTION__, ys->ys_argument, type, 
 		 *yrestype?(*yrestype)->ys_argument:"null");